Adding some error messages for LInf and Max templates.  I implemented an analytical gradient function for these two metrics, so that the code would not just default to the numerical gradient.  These new functions just set an error explaining the problem.  If the user wants to do this anyway, he/she can explicitly set the gradient type of the OF.
We probably still need to add a warning for when the user does select the
numerical gradient, but I believe this is lower priority.

The Hessians for these cases are already handled.  We don't allow numerical
Hessians, so you can only create one if the Analytic version is defined.

The modification to ConjugateGradient is just so that the error message is passed
back to the user correctly when this case happens (LInf or Max template with
CG as the solver).

Passes Unit tests.