New functions ZZ *Zoltan_Copy(ZZ *from) int Zoltan_Copy_To(ZZ *to, ZZ *from)
The first creates and returns a new Zoltan_Struct which is a copy
of the argument.  The second copies one Zoltan_Struct to another.
These functions have been requested in the past, and they will also
serve to support a copy operator and copy constructor in the C++
API to Zoltan.

I also did some reorganization in the file so the new functions fit
in better.