Changed the hypergraph interface functions to allow Zoltan to ignore (and not allocate space for) very large edges. Also added GIDs and LIDs for edges.
Previously, the application returned the total number of pins, and
then returned a count and list of vertices for each edge.

Now, the application returns the count of vertices for each edge in
a separate function.  The counts can be processed before edge information
is requested.  This change allows Zoltan to not request a list
of vertices for dense edges.