Added the SWIG interface (Amesos.i) to wrap serial Amesos classes into the python word. All I did is largely inspired from the python/epetra interface. Many thanks to Bill Spotz for his very nice work.
Amesos/python lives in the PyTrilinos package, and allows the user to build up
an Epetra_CrsMatrix and Epetra_MultiVectors, an Epetra_LinearProblem, and finally
access all the enabled classes of Amesos to solve the corresponding linear
problem. The differences with respect to the C++ codes are:
- only serial at this point;
- since I had some problems to wrap the parameter list, the amesos/python
  does not have SetParameters(), instead it uses SetInt(), SetBool(),
  SetDouble() and SetString() to toggle the behavior of the undelying class.

Two examples are available in the amesos/python/example subdirectory,
showing how to use the factory and how to build a concrete class without
the factory.

I have tested the code with LAPACK, KLU and UMFPACK.