This is a major commit that allows LOCA to solve constrained nonlinear systems and do continuation of constrained systems.  Continuation is just a special case of a constrained nonlinear system, so the continuation groups have been rewritten to use this functionality.  The predictor interface has also been changed to support higher-order predictors.  Finally, the LOCA Factory interface has been reworked slightly to be more flexible when creating strategies.
Most of the constraint code seems to work, although there appears to be an
issue with the scaling of constraint equations when using arc-length scaling.

The LOCA interface is currently evolving to bring the single parameter and
multi-parameter continuation code into a single continuation interface.
Right now things are a little messy because both the old and new code is
in the repository, but once the multiparameter code is completed we will be
able to get rid of all of the old single parameter code.

The MF interface is still broken, so don't enable MF.