Modified results files to allow for more sane parsing once they reach the TrilinosTestResults repository.
Changed log file setup so that all runconfigure and runmake log output for one build all goes to the same file...novel idea.

Added config.log contents to results files for failed configure attempts.

Added --simple-dir-name option to runtests for use by runharness so that test directory name isn't timestamped.  This will make it easier to find the test results directory when parsing results.  There is no need to serialize the test directory name when in the context of a single build.