I have added the make targets runtests-serial and runtests-mpi to the makefiles for all of Trilinos and specifically for the pacakges teuchos, thyra and epetra.  I have also added the epetra/thyra adapter testers to the epetra/test/defintion file.  This was the only way that I know of to get the new test suite to run the epetra/thyra adapter tests.  I do not know what will happen when the epetra/thyra adapters are not enabled (i.e. when --enable-epetra-thyra is not used at configure time).
To run the test suite (which calls the new test harness) just do a make first and then type:

  make runtests-serial

and you will run the entire test suite (for the base Trilinos test directory).

To run the test suite for just teuchos, thyra or epetra, just do a make as usuall and then cd to pacakge/test (where package = teuchos, thyra or epetra) and then type:

  make runtests-serial

I have not tested the runtests-mpi target yet but it expects to you supply the variable TRILINOS_MPI_GO (either in the environment or as an argument to make) which is then passed to the --mpi-go option.