Significant change to timing facility.  Previous version assumed all processors in a communicator would start/stop timer the same number of times. This assumption was correct for our coarse-grain timing, but is not correct for fine-grain timing.
Synchronization of timers within Zoltan_Timer_Stop was removed and moved
to Zoltan_Timer_Print.  The assumption now is that all processors will
call Zoltan_Timer_Print the same number of times -- an assumption that
should be easier to live with.

HERE IS THE BIGGEST DIFFERENCE:
The global time reported used to be
    sum_over_time_steps ( max_over_processors (time_for_one_time_step))

It is now
    max_over_processors ( sum_over_time_steps (time_for_one_time_step))

I preferred the former, but it required that all processors have the same
number of "time steps."

C++ interface is not quite right yet.