Minor timing output and orthogonalization output formatting changes.
Fixed computation of residual in LOBPCG to use the 2-norm instead of norm
defined by the problem.  This discrepency was causing confusion about the
correctness of the eigenvectors returned from the solver.
Now the generalized eigenproblem examples work correctly for LOBPCG.
There's still an issue with the standard eigenvalue problem, inserting an
identity mass matrix resolves the issue.  It's not right, but it will help
me figure out my bug :)